+++ query/src/cuda/scan_utils.h
@@ -5,5 +5,5 @@
 // Performs exclusive scan (prefix sum) on device array
 // Returns total sum
 extern "C" {
-    int exclusive_scan_gpu(const int* d_input, int* d_output, int num_elements);
+    long long exclusive_scan_gpu(const int* d_input, long long* d_output, int num_elements);
 }
